requirements.txt
numpy
opencv-python
Pillow
scipy
scikit-learn
pyyaml
torch
torchvision